test(settings): migrate settings tests across category files
Restore full green after the P1-C3/C4 UI rework (test surface now
resolves to the compact list -> category detail two-step flow).

- Add shared helpers: buildSettingsApp({platform, locale,
  size = 390x844}) with view port control, openCategory, and
  scrollUntilFound resolved by Key('settingsScrollArea') (DR-13).
- Migrate the 22 legacy cases into 5 files: settings_navigation_test
  (title, 4 categories, push/back, ja names), category_display_test,
  category_behavior_test (incl. CJK platform gates, wheel note with an
  inline ProviderScope override, keep-keyboard Android/iOS),
  category_connection_test (drops the IMAGE TRANSFER uppercase finder
  for the new Image Transfer header; Clear Host Keys confirm/cancel
  now expects the l10n snackbar), category_about_test.
- Add the S-6 scroll-position retention assertion: a short viewport
  forces the list to scroll, then push -> back must restore the same
  offset.
- Delete test/screens/settings_screen_test.dart.
- Fix settings_category_list to a Column layout so it can live inside
  the compact SliverList without nesting a second viewport.
